fea: prevent listen before set shopId

+++ src/components/products/updates_listener.ts
@@ -4,52 +4,53 @@ import { Product, Barcode } from '../products/model'
 
 async function listen() {
 	const meta = await Meta.findOne({ attributes: ['shopId'] })
+	if (meta) {
+		/**
+			Products 
+		*/
+		firebaseConnection
+			.collection('updates')
+			.where('shopId', '==', meta!.shopId)
+			.where('table', '==', 'products')
+			.onSnapshot(snap => {
+				snap.docs.forEach(async doc => {
+					const { data, type } = doc.data()
 
-	/**
-		Products 
-	*/
-	firebaseConnection
-		.collection('updates')
-		.where('shopId', '==', meta!.shopId)
-		.where('table', '==', 'products')
-		.onSnapshot(snap => {
-			snap.docs.forEach(async doc => {
-				const { data, type } = doc.data()
+					if (type == 'create') {
+						await Product.create(data, { include: { model: Barcode, as: 'barcodes' } })
+					} else if (type == 'update') {
+						await Product.update(data, { where: { id: data.id } })
+					}
 
-				if (type == 'create') {
-					await Product.create(data, { include: { model: Barcode, as: 'barcodes' } })
-				} else if (type == 'update'){
-					await Product.update(data, { where: { id: data.id } })
-				}
-
-				firebaseConnection.collection('updates').doc(doc.id).delete()
+					firebaseConnection.collection('updates').doc(doc.id).delete()
+				})
+			}, error => {
+				throw error
 			})
-		}, error => {
-			throw error
-		})
 
-	/**
-		Barcodes
-	*/
-	firebaseConnection
-		.collection('updates')
-		.where('shopId', '==', meta!.shopId)
-		.where('table', '==', 'barcodes')
-		.onSnapshot(snap => {
-			snap.docs.forEach(async doc => {
-				const { data, type } = doc.data()
+		/**
+			Barcodes
+		*/
+		firebaseConnection
+			.collection('updates')
+			.where('shopId', '==', meta!.shopId)
+			.where('table', '==', 'barcodes')
+			.onSnapshot(snap => {
+				snap.docs.forEach(async doc => {
+					const { data, type } = doc.data()
 
-				if (type == 'create') {
-					await Barcode.create(data)
-				} else if (type == 'update'){
-					await Barcode.update(data, { where: { id: data.id } })
-				}
+					if (type == 'create') {
+						await Barcode.create(data)
+					} else if (type == 'update') {
+						await Barcode.update(data, { where: { id: data.id } })
+					}
 
-				firebaseConnection.collection('updates').doc(doc.id).delete()
+					firebaseConnection.collection('updates').doc(doc.id).delete()
+				})
+			}, error => {
+				throw error
 			})
-		}, error => {
-			throw error
-		})
+	}
 }
 
 export { listen }
